Refactor the captcha component for reusability
